# Improve lexer perf for common cases