--- a/.gitignore
+++ b/.gitignore
@@ -1,2 +1,4 @@
 jpgo
 jmespath-fuzz.zip
+cpu.out
+go-jmespath.test